refactor(k3s-vms): Managing all resources with the module.
Next thing I will be working towards is getting them in a for_each
so things are dynamic. I was having some troubles earlier, so decided to
do it this way for now.

The issue I was facing was with the import. It seemed like whenever I would import
the first vm, all would be good, but upon importing the second one, the data was mixed up.
Could it be that there is something wrong with the bpg provider and importing data?